refactor(insights): share the toolbar's tuned element capture
The copy-image action called html-to-image directly and missed the style-property list the toolbar passes. Without that list the library copies every computed property of the document element onto every cloned node, and most of them are design-token custom properties that resolve into the copied values anyway.

Moves the cached, custom-property-free list and the toBlob wrapper into lib/utils/captureElementImage.ts. The toolbar screenshot helpers and the copy-image action now both go through it.

Generated-By: PostHog Desktop
Task-Id: 74d5a998-1b5e-49eb-a719-202ab137fce2
