+++ frontend/src/toolbar/utils/screenshot.ts
@@ -1,4 +1,4 @@
-import { toBlob } from 'html-to-image'
+import { captureElementImage } from 'lib/utils/captureElementImage'
 
 import { toolbarUploadMedia } from '~/toolbar/toolbarFetch'
 import { TOOLBAR_ID } from '~/toolbar/utils'
@@ -11,24 +11,6 @@ function screenshotFilter(node: Node): boolean {
     return !(node instanceof HTMLElement && node.id === TOOLBAR_ID)
 }
 
-let cachedStylePropertyNames: string[] | null = null
-
-const getAllStylePropertyNames = (): string[] => {
-    if (cachedStylePropertyNames) {
-        return cachedStylePropertyNames
-    }
-    const names: string[] = []
-    const style = getComputedStyle(document.documentElement)
-    for (let i = 0; i < style.length; i++) {
-        const name = style[i]
-        if (!name.startsWith('--')) {
-            names.push(name)
-        }
-    }
-    cachedStylePropertyNames = names
-    return names
-}
-
 export interface CaptureOptions {
     pixelRatio?: number
     width?: number
@@ -37,19 +19,12 @@ export interface CaptureOptions {
 }
 
 export async function captureElementScreenshot(element: HTMLElement, options?: CaptureOptions): Promise<Blob> {
-    const blob = await toBlob(element, {
+    return captureElementImage(element, {
         type: 'image/jpeg',
-        includeStyleProperties: getAllStylePropertyNames(),
         quality: 0.7,
         filter: screenshotFilter,
         ...options,
     })
-
-    if (!blob) {
-        throw new Error('Failed to capture element screenshot')
-    }
-
-    return blob
 }
 
 export async function uploadScreenshot(blob: Blob): Promise<ElementScreenshot> {
